refactor(mobile): rename account/[account] to account/[accountId]
