+++ apps/mobile/src/app/account/[accountId]/collectibles.tsx
