+++ apps/mobile/src/features/account/accounts-widget.tsx
@@ -96,7 +96,6 @@ export function AccountsWidget() {
                     router.navigate({
                       pathname: AppRoutes.Account,
                       params: {
-                        account: account.id,
                         accountId: account.id,
                       },
                     });
